requirements.txt
Flask==3.0.3
Authlib==1.3.1
PyMySQL==1.1.1
sseclient-py==1.8.0
python-dotenv==1.0.1